sparse: add SparseSolverBaseVisitor
